requirements.txt
pytest>=7.4
pytest-playwright>=0.4
httpx>=0.25
python-dotenv>=1.0
